Make sure postinstall does not build in development environment.
Related to pull request !10. 10 fixes most problems but there
is one small caveat.

Since we have added 'build' target in 'postinstall', everytime
we install dependencies by running 'npm install'. The 'postinstall'
target will be run and the project will get built. That is not
something we may necessarily want in `dev` environment.

So fix that issue we are using `per-env` package we checks the
environment and skips build in `dev` but not in `prod` env.